services/budapp/requirements.txt
# Service/Clients
dapr >= 1.13.0

# API
fastapi >= 0.111.1
uvicorn >= 0.30.3
pydantic >= 2.8.2
pydantic-settings >= 2.3.4
email_validator>=2.2.0
aiohttp>=3.10.5
python-multipart>=0.0.17
aiofiles>=24.1.0

# Utils
structlog >= 24.4.0
cachetools >= 5.4.0
ujson >= 5.10.0
PyYAML>=6.0.2
huggingface-hub>=0.26.5
datasets
lxml

# Database
SQLAlchemy>=2.0.36
sqlalchemy-utils>=0.41.2
psycopg
psycopg2-binary>=2.9.10
alembic>=1.13.3
alembic-postgresql-enum>=1.3.0

# Security
passlib>=1.7.4
bcrypt>=4.2.0
python-jose[cryptography]>=3.3.0

# Redis
redis>=5.2.1

# Service
budmicroframe @ git+https://github.com/BudEcosystem/bud-microframe.git

# Keycloak
python-keycloak>=5.3.1
# Minio
minio>=7.2.14

# BeautifulSoup
beautifulsoup4>=4.13.4

# PyPDF2
PyPDF2>=3.0.1
